Assert gymnasium API conformance via SB3's env_checker
Validates the contract directly for all four registered environments,
including the ones with no golden fixtures.

check_env caught a real declaration bug in cartpole and quadrotor: both
_set_observation_space methods declared Box(..., dtype=np.float32) for
state_space/observation_space, but both envs actually emit float64
observations (via np.hstack over PyBullet joint-state values), the same
pattern inverted_pendulum.py already declares float64 for. Corrected the
declared dtype to float64 in both files to match emitted data -- no
observation values were cast, clipped, or otherwise altered.

+++ safe_control_gym/envs/gym_control/cartpole.py
@@ -497,7 +497,7 @@ class CartPole(BenchmarkEnv):
                               self.x_dot_threshold,
                               self.theta_threshold_radians * 2,
                               self.theta_dot_threshold])
-        self.state_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float32)
+        self.state_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float64)
 
         # Concatenate goal info for RL
         if self.COST == Cost.RL_REWARD and self.TASK == Task.TRAJ_TRACKING and self.obs_goal_horizon > 0:
@@ -509,7 +509,7 @@ class CartPole(BenchmarkEnv):
             obs_bound = np.concatenate([obs_bound] * 2)
         # Define obs space exposed to the controller
         # Note obs space is often different to state space for RL (with additional task info)
-        self.observation_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float32)
+        self.observation_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float64)
 
         # Define obs/state labels and units.
         self.STATE_LABELS = ['x', 'x_dot', 'theta', 'theta_dot']
